requirements.txt
Django
djangorestframework
flake8
psycopg2
requests
celery
django-smtp-ssl
amqp==2.5.1
django-rest-swagger
django-cors-headers
kombu==4.6.5
django-model-utils
django-autoslug
